Approval: move get_or_create_from_valid() to User model
